trying shadow dom method
